Rename packages, separate AE logger
