+++ packages/scandipwa/packages/after-emit-logger/build-config/config.plugin.js
@@ -1,0 +1,11 @@
+const WebpackAfterEmitLogger = require('./webpack-after-emit-logger');
+
+module.exports = {
+    plugin: {
+        overrideWebpackConfig: ({ webpackConfig }) => {
+            webpackConfig.plugins.push(new WebpackAfterEmitLogger);
+
+            return webpackConfig;
+        }
+    }
+};
