+++ packages/scandipwa/packages/after-emit-logger/build-config/webpack-after-emit-logger.js
@@ -1,0 +1,15 @@
+const afterEmitLogger = require('../src');
+
+class WebpackAfterEmitLogger {
+    pluginMeta = { name: 'WebpackAfterEmitLogger' };
+
+    apply(compiler) {
+        // TODO check if it works
+        compiler.hooks.afterEmit.tap(
+            this.pluginMeta,
+            () => afterEmitLogger.emitLogs()
+        );
+    }
+}
+
+module.exports = WebpackAfterEmitLogger;
