+++ packages/scandipwa/packages/after-emit-logger/package.json
@@ -1,0 +1,15 @@
+{
+    "name": "@scandipwa/after-emit-logger",
+    "version": "0.0.0",
+    "license": "proprietary",
+    "main": "src/index.js",
+    "dependencies": {
+        "@scandipwa/scandipwa": "3.2.4-alpha.2"
+    },
+    "scandipwa": {
+        "type": "extension"
+    },
+    "eslintConfig": {
+        "extends": "@scandipwa"
+    }
+}
