+++ packages/scandipwa/packages/runtime-i18n/build-config/config.plugin.js
