+++ packages/scandipwa/packages/runtime-i18n/build-config/webpack-i18n-import-loader/index.js
