+++ packages/scandipwa/packages/runtime-i18n/src/plugin/App.plugin.js
