+++ packages/scandipwa/packages/runtime-i18n/src/util/mergeTranslations.js
