--- packages/scandipwa/packages/scandipwa-runtime-i18n/build-config/webpack-i18n-tracker/after-emit-logger/index.js
@@ -1,26 +1,0 @@
-const logger = require('@scandipwa/scandipwa-dev-utils/logger');
-
-function clearArray(arr) {
-    arr.splice(0, A.length);
-}
-
-class Logger {
-    persistedLogs = [];
-    runtimeLogs = [];
-
-    addRuntimeLog(log) {
-        this.runtimeLogs.push(log);
-    }
-
-    emitLogs() {
-        const emitLog = ({ type, args }) => logger[type](...args);
-
-        setTimeout(() => {
-            this.persistedLogs.forEach(emitLog);
-            this.runtimeLogs.forEach(emitLog);
-            clearArray(this.runtimeLogs);
-        }, 100);
-    }
-}
-
-module.exports = new Logger;
